Refactor: spembedded.ts convert to module and name change
