+++ src/components/files/filesTypes.ts
@@ -1,9 +1,6 @@
 import { DriveItem } from "@microsoft/microsoft-graph-types-beta";
-import {
-  IArchiveClientProgress,
-  IContainer,
-} from "../../common/types";
-import { IJobProgress } from "../../services/spembedded";
+import { IArchiveClientProgress, IContainer } from "../../common/types";
+import { IJobProgress } from "../../services/backendApi";
 
 /**
  * Files 入口组件属性。
